Breakout DrawableFactory
